requirements-gpu.txt
absl-py==0.1.13
astor==0.6.2
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
boto==2.38.0
chardet==2.3.0
cloud-init==17.2
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
cycler==0.10.0
defer==1.0.6
gast==0.2.0
google-compute-engine==2.7.5
graphviz==0.8.2
grpcio==1.10.1
h5py==2.7.1
html5lib==0.9999999
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
Keras==2.1.5
kiwisolver==1.0.1
language-selector==0.1
lxml==3.5.0
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
numpy==1.14.2
oauthlib==1.0.3
opencv-python==3.4.0.12
Pillow==5.1.0
pkg-resources==0.0.0
prettytable==0.7.2
protobuf==3.5.2.post1
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pydot==1.2.4
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.2
python-debian==0.1.27
python-systemd==231
pytz==2018.4
pyxdg==0.25
PyYAML==3.11
requests==2.9.1
scipy==1.0.1
screen-resolution-extra==0.0.0
six==1.10.0
ssh-import-id==5.5
system-service==0.3
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
Werkzeug==0.14.1
xkit==0.0.0
